# PERF: check first for recognized scalars in DatetimeArray._validate_scalar