--- a/pandas/core/arrays/datetimelike.py
+++ b/pandas/core/arrays/datetimelike.py
@@ -635,8 +635,8 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
         -------
         self._scalar_type or NaT
         """
-        if isinstance(value, self._recognized_scalars):
-            value = self._scalar_type(value)
+        if isinstance(value, self._scalar_type):
+            pass
 
         elif isinstance(value, str):
             # NB: Careful about tzawareness
@@ -655,6 +655,9 @@ class DatetimeLikeArrayMixin(OpsMixin, NDArrayBackedExtensionArray):
             #  or else we'll fail to raise in _unbox_scalar
             msg = self._validation_error_message(value, allow_listlike)
             raise TypeError(msg)
+
+        elif isinstance(value, self._recognized_scalars):
+            value = self._scalar_type(value)
 
         else:
             msg = self._validation_error_message(value, allow_listlike)