Something small, but most of the time we will be setting with actual scalars, so putting that first improves this case a bit:

```python
arr = pd.date_range("2012-01-01", periods=3).array
scalar = arr[0]

In [3]: %timeit arr._validate_setitem_value(scalar)
3.04 µs ± 58.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- master
2.08 µs ± 27.2 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- PR
```

(it gives a noticeable difference in eg fillna benchmarks)