@jorisvandenbossche 2021-11-24
The second commits gives another small improvement (the `_assert_tzawareness_compat` just before already ensures that self and other have either both a tz or both not):

```python
In [3]: %timeit arr._validate_setitem_value(scalar)
3.04 µs ± 58.5 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- master
1.85 µs ± 82 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)  # <-- pR
```