@jreback 2021-11-25
thanks. at some point if you can add the collected PR numbers to the whatsnew in the perf improvements section.