@jorisvandenbossche 2021-11-25
@jreback they're mostly related to (or most relevant for) ArrayManager, so I should probably start with some actual basic docs about it before starting to mention it in the whatsnew