@jreback 2021-11-25
well u can certainly make a section in the whatsnew to indicate that perf for these have changed